Adds openshift engine
