+++ container/docker/templates/conductor-dockerfile.j2
@@ -18,6 +18,8 @@ RUN apk add --no-cache -U python2-dev git rsync libffi libffi-dev openssl openss
 
 ADD https://get.docker.com/builds/Linux/x86_64/docker-{{ docker_version }}.tgz /tmp/docker.tgz
 ADD https://raw.githubusercontent.com/ansible/ansible/devel/lib/ansible/module_utils/k8s_common.py /usr/lib/python2.7/site-packages/ansible/module_utils/k8s_common.py
+ADD https://raw.githubusercontent.com/ansible/ansible/devel/lib/ansible/module_utils/openshift_common.py /usr/lib/python2.7/site-packages/ansible/module_utils/openshift_common.py
+
 COPY /contrib/get-pip.py /get-pip.py
 RUN python /get-pip.py && \
     mkdir -p /etc/ansible/roles /_ansible/src && \
